refactor: seperate business logic from ui logic
